requirements-test.txt
# Requirements for executing tests in test environments
coverage
mock>=1.0.1
nose>=1.3.0
git+git://github.com/django-nose/django-nose.git#egg=django-nose

# Required to test default models
django-extensions>=0.9

# Required for mocking signals
mock-django>=0.6.5
